get ready for new GroupFromSQL release

+++ RM -Group from SQL/GroupFromSQL.py
@@ -1,10 +1,11 @@
-import os
 import sys
-from datetime import datetime
+sys.path.append(r'..\RM -RMpy package')
+import RMpy.launcher  # type: ignore
+import RMpy.common as RMc  # type: ignore
+from RMpy.common import q_str # type: ignore
 
-sys.path.append( r'..\\RM -RMpy package' )
-import RMpy.launcher # type: ignore
-import RMpy.common as RMc # type: ignore
+import os
+
 
 # Always make a database backup before using this script.
 
@@ -12,17 +13,17 @@ import RMpy.common as RMc # type: ignore
 #   RootsMagic database file
 #   RM-Python-config.ini
 
-# Tested with: 
-#   RootsMagic database file v10.0.1
-#   Python for Windows v3.12.3
+# Tested with:
+#   RootsMagic database file v10
+#   Python for Windows v3.12
 
 # Config files fields used
 #    [FILE_PATHS]  REPORT_FILE_PATH
-#    [FILE_PATHS]   REPORT_FILE_DISPLAY_APP
+#    [FILE_PATHS]  REPORT_FILE_DISPLAY_APP
 #    [FILE_PATHS]  DB_PATH
 #    [FILE_PATHS]  RMNOCASE_PATH
-#
 #    [OPTIONS]     GROUP_NAME
+#  note: GROUP_NAME may be a list of names, one per line.
 
 #  note:Section name is the same label as [OPTIONS] GROUP_NAME's value
 #    [GROUP_NAME_Value]  SQL_QUERY   the SQL to select group members
@@ -38,11 +39,11 @@ def main():
     RegExp_required = False
 
     RMpy.launcher.launcher(os.path.dirname(__file__),
-                    config_file_name,
-                    run_selected_features,
-                    allow_db_changes,
-                    RMNOCASE_required,
-                    RegExp_required )
+                           config_file_name,
+                           run_selected_features,
+                           allow_db_changes,
+                           RMNOCASE_required,
+                           RegExp_required)
 
 
 # ===================================================DIV60==
@@ -55,21 +56,43 @@ def run_selected_features(config, db_connection, report_file):
 def RunSQLGroupFeature(config, db_connection, report_file):
 
     try:
-        group_name = config['OPTIONS']['GROUP_NAME']
+        group_name_list = config['OPTIONS'].get(
+            'GROUP_NAME').split('\n')
     except:
         raise RMc.RM_Py_Exception(
             'section: [OPTIONS],  key: GROUP_NAME   not found.')
-    try:
-        config[group_name]
-    except:
-        raise RMc.RM_Py_Exception(
-            'section: [' + group_name + ']   not found.')
-    try:
-        SQL_statement =config[group_name]['SQL_QUERY']
-    except:
-        raise RMc.RM_Py_Exception(
-            'section: [' + group_name + '],  key: SQL_QUERY   not found.')
 
+    for group_name in group_name_list:
+        if group_name == '':
+            continue
+        try:
+            config[group_name]
+        except:
+            raise RMc.RM_Py_Exception(
+                f'section: [{q_str(group_name)}]   not found.')
+
+    for group_name in group_name_list:
+        if group_name == '':
+            continue
+        update_group(db_connection, config, report_file, group_name)
+
+    return
+
+
+# ===================================================DIV60==
+def update_group(db_connection, config, report_file, group_name):
+
+    GroupID = confirm_DB_group_name(group_name, report_file, db_connection)
+
+    try:
+        SQL_statement = config[group_name]['SQL_QUERY']
+    except:
+        raise RMc.RM_Py_Exception(
+            f'section: [{q_str(group_name)}],  key: SQL_QUERY not found.')
+
+    if SQL_statement == '':
+        raise RMc.RM_Py_Exception(
+            'section: [GROUP_NAME], key: SQL_QUERY No value entered. Nothing to do.')
 
     viewStmt = "DROP VIEW IF EXISTS PersonIdList_RJO_utils"
     try:
@@ -77,19 +100,7 @@ def RunSQLGroupFeature(config, db_connection, report_file):
         cur.execute(viewStmt)
     except RMc.RM_Py_Exception as e:
         # database open issues don't appear until here !
-        raise RMc.RM_Py_Exception(
-            'Cannot open the database.\n' + str(e))
-
-    if group_name == '':
-        raise RMc.RM_Py_Exception(
-            'section: [OPTIONS], key: group_name \n' +
-            'No value entered.\n Nothing to do.')
-    
-    if SQL_statement == '':
-        raise RMc.RM_Py_Exception(
-            'section: [GROUP_NAME], key: SQL_QUERY \n' +
-            'No value entered.\n Nothing to do.')
-
+        raise RMc.RM_Py_Exception(f'Cannot open the database.\n{str(e)}')
 
     # generate the SQL statement and create the view
     SqlStmt = "CREATE TEMP VIEW PersonIdList_RJO_utils AS " + SQL_statement
@@ -107,30 +118,30 @@ def RunSQLGroupFeature(config, db_connection, report_file):
         cur = db_connection.cursor()
         cur.execute(SqlStmt)
         numInView = cur.fetchone()[0]
-        report_file.write("# of persons selected: " + str(numInView) + "\n")
+        report_file.write(
+            str(numInView) + " persons selected by the SQL statement:\n")
     except:
         raise ("ERROR: SQL_QUERY returned an error when run as a VIEW. \n\n" +
                "SQL entered was:\n" + SQL_statement + "\n")
 
-    GroupID = ConfirmGroup(group_name, report_file, db_connection)
+    report_file.write(SQL_statement + "\n\n\n")
 
     PopulateGroup(GroupID, db_connection)
-
-    report_file.write("\nSQL used to specify group membership:\n\n" + SQL_statement)
 
     viewStmt = "DROP VIEW IF EXISTS PersonIdList_RJO_utils"
     cur = db_connection.cursor()
     cur.execute(viewStmt)
 
-    return
-
 
 # ===================================================DIV60==
-def ConfirmGroup(Name, report_file, db_connection):
+def confirm_DB_group_name(Name, report_file, db_connection):
     #  TagTable
     #   TagID=rowid
     #   TagType =0 for Groups
     #   TagName   duplicates not constrained
+
+    Divider = "="*50 + "===DIV60=="
+    report_file.write(f"{Divider}\n\n")
 
     # check how many groupNames with name and TagTape=0 already exist
     SqlStmt = """
@@ -145,16 +156,15 @@ WHERE TagName=? COLLATE NOCASE AND TagType=0 COLLATE NOCASE
     GroupID = result[1]
 
     if existingNumber > 1:
-        raise RMc.RM_Py_Exception("\nERROR: Group: " + Name +
-              " already exists more than once.\n Use a different name, or rename one of them. \n")
+        raise RMc.RM_Py_Exception(f"\nERROR: Group: {q_str(Name)}  exists more than once in the database.\n"
+                                  "Rename one of the duplicates. \n")
 
     if existingNumber == 1:
-        report_file.write("\nINFO: Group: " + Name +
-              " exists and will be updated. \n")
+        report_file.write(f"Group: {q_str(Name)} will be updated. \n")
 
     else:  # existingNumber == 0
-        raise RMc.RM_Py_Exception("\nINFO: Group: " 
-            + Name + " does not exist in the database. \n")
+        raise RMc.RM_Py_Exception(
+            f"\nERROR: Group: {q_str(Name)} does not exist in the database. \n")
 
     return GroupID
 
@@ -171,21 +181,20 @@ def PopulateGroup(GroupID, dbConnection):
     try:
         cur = dbConnection.cursor()
         cur.execute(SqlStmt, (GroupID,))
-    except RMc.RM_Py_Exception: (
-        'Cannot clear the group members. Close RM and try again.\n' + str(e))
+    except RMc.RM_Py_Exception:
+        ('Cannot clear the group members. Close RM and try again.\n' + str(e))
 
     # add the members
     SqlStmt = """
-  INSERT INTO GroupTable
-  SELECT
-   null
+INSERT INTO GroupTable
+SELECT
+    null
    ,?        AS GroupID
    ,PersonID AS StartID
    ,PersonID AS EndID
    ,(julianday('now') - 2415018.5) AS UTCModDate 
-
-  FROM PersonIdList_RJO_utils
-  """
+FROM PersonIdList_RJO_utils
+"""
     try:
         cur = dbConnection.cursor()
         cur.execute(SqlStmt, (GroupID,))
